# backend/internal/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/asset": {
            "get": {
                "tags": [
                    "asset"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page_index",
                        "name": "page_index",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page_size",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name or ip",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "asset id",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "asset ids",
                        "name": "ids",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "asset's parent id",
                        "name": "parent_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "asset name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "asset ip",
                        "name": "ip",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "is info mode",
                        "name": "info",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.HttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/controller.ListData"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "list": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/model.Asset"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "asset"
                ],
                "parameters": [
                    {
                        "description": "asset",
                        "name": "asset",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Asset"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.HttpResponse"
                        }
                    }
                }
            }
        },
        "/asset/:id": {
            "put": {
                "tags": [
                    "asset"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "asset id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "asset",
                        "name": "asset",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Asset"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.HttpResponse"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "asset"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "asset id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.HttpResponse"
                        }
                    }
                }
            }
        },
        "/asset/:id/permissions": {
            "get": {
                "tags": [
                    "asset"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "asset id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "account ids (comma separated, e.g. 123,456,789)",
                        "name": "account_ids",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.HttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/controller.AssetPermissionMultiAccountResult"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controller.AssetPermissionBatchResult": {
            "type": "object",
            "properties": {
                "results": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/controller.AssetPermissionResult"
                    }
                }
            }
        },
        "controller.AssetPermissionMultiAccountResult": {
            "type": "object",
            "properties": {
                "results": {
                    "description": "accountId -\u003e batch results",
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/controller.AssetPermissionBatchResult"
                    }
                }
            }
        },
        "controller.AssetPermissionResult": {
            "type": "object",
            "properties": {
                "allowed": {
                    "type": "boolean"
                },
                "reason": {
                    "type": "string"
                },
                "restrictions": {
                    "type": "object",
                    "additionalProperties": true
                },
                "rule_id": {
                    "type": "integer"
                },
                "rule_name": {
                    "type": "string"
                }
            }
        },
...
        "controller.HttpResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "controller.ListData": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "list": {
                    "type": "array",
                    "items": {}
                }
            }
        },
        "model.AccessAuth": {
            "type": "object",
            "properties": {
                "allow": {
                    "type": "boolean"
                },
                "cmd_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "end": {
                    "type": "string"
                },
                "ranges": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Range"
                    }
                },
                "start": {
                    "type": "string"
                }
            }
        },
...
        "model.AccessTimeControl": {
            "type": "object",
            "properties": {
                "comment": {
                    "description": "Description of the time restriction",
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "time_ranges": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.TimeRange"
                    }
                },
                "timezone": {
                    "description": "e.g., \"Asia/Shanghai\"",
                    "type": "string"
                }
            }
        },
...
        "model.AccountAuthorization": {
            "type": "object",
            "properties": {
                "permissions": {
                    "description": "V2 permissions (connect, file_upload, etc.)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.AuthPermissions"
                        }
                    ]
                },
                "rids": {
                    "description": "Role IDs for ACL system",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "model.Asset": {
            "type": "object",
            "properties": {
                "access_auth": {
                    "description": "Deprecated: Use V2 fields below",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.AccessAuth"
                        }
                    ]
                },
                "access_time_control": {
                    "description": "V2 Access Control (replaces AccessAuth)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.AccessTimeControl"
                        }
                    ]
                },
                "asset_command_control": {
                    "$ref": "#/definitions/model.AssetCommandControl"
                },
                "authorization": {
                    "$ref": "#/definitions/model.AuthorizationMap"
                },
                "comment": {
                    "type": "string"
                },
                "connectable": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "creator_id": {
                    "type": "integer"
                },
                "gateway_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "ip": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "node_chain": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "integer"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "protocols": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "resource_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "updater_id": {
                    "type": "integer"
                }
            }
        },
        "model.AssetCommandControl": {
            "type": "object",
            "properties": {
                "cmd_ids": {
                    "description": "Command IDs to control",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "comment": {
                    "description": "Description of the command restriction",
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "template_ids": {
                    "description": "Command template IDs",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "model.AuthPermissions": {
            "type": "object",
            "properties": {
                "connect": {
                    "type": "boolean"
                },
                "copy": {
                    "type": "boolean"
                },
                "file_download": {
                    "type": "boolean"
                },
                "file_upload": {
                    "type": "boolean"
                },
                "paste": {
                    "type": "boolean"
                },
                "share": {
                    "type": "boolean"
                }
            }
        },
...
        "model.AuthorizationMap": {
            "type": "object",
            "additionalProperties": {
                "$ref": "#/definitions/model.AccountAuthorization"
            }
        },
...
        "model.Range": {
            "type": "object",
            "properties": {
                "times": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "week": {
                    "type": "integer"
                }
            }
        },
...
        "model.TimeRange": {
            "type": "object",
            "properties": {
                "end_time": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "weekdays": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
    }
}